update Modal components
