+++ src/components/modal/components/content.js
@@ -1,36 +1,37 @@
-import React from 'react';
+import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
 
-const ModalContent = ({
-  children,
-  className,
-  style,
-  renderAs,
-}) => {
-  const Element = renderAs;
-  return (
-    <Element
-      style={style}
-      className={classnames('modal-content', className)}
-    >
-      {children}
-    </Element>
-  );
-};
+export default class ModalContent extends PureComponent {
+  static propTypes = {
+    children: PropTypes.node,
+    className: PropTypes.string,
+    style: PropTypes.object,
+    renderAs: PropTypes.string,
+  }
 
-ModalContent.propTypes = {
-  children: PropTypes.node,
-  className: PropTypes.string,
-  style: PropTypes.object,
-  renderAs: PropTypes.string,
-};
+  static defaultProps = {
+    children: null,
+    className: '',
+    style: {},
+    renderAs: 'div',
+  }
 
-ModalContent.defaultProps = {
-  children: null,
-  className: '',
-  style: {},
-  renderAs: 'div',
-};
-
-export default ModalContent;
+  render() {
+    const {
+      children,
+      className,
+      style,
+      renderAs,
+    } = this.props;
+    const Element = renderAs;
+    return (
+      <Element
+        style={style}
+        className={classnames('modal-content', className)}
+      >
+        {children}
+      </Element>
+    );
+  }
+}
